fix: refetch invoices after mutation
if a filter is selected, then there are too many combinations of filters to be able to effectively
optimistically update all of them with the new/updated data. Instead, invalidate all the list
queries so they refecth. We can still optimistically update the non-filtered list, and the
individual detail page, so in many cases, the optimistic behaviour is preserved